requirements.txt
torch
torchaudio
transformers
accelerate
soundfile
pyaudio
numpy
groq
python-dotenv
